Support for SEP Scoreboard URLs.
Added new file scoreboard_call which contains the
function to generate custom SEP Scoreboard URLs. All
that function needs is an input time (probably the time
that the report is generated/emailed out), a list of
models, and a toggle for probability/intensity.

Added a couple of lines of text to the email to include
the custom URLs.